changed definition of Index from short to long because value may exceed capacity of short

+++ kdebindings/smoke/smoke.h
@@ -83,7 +83,7 @@ public:
 	EnumToLong
     };
 
-    typedef short Index;
+    typedef long Index;
     typedef void (*ClassFn)(Index method, void* obj, Stack args);
     typedef void* (*CastFn)(void* obj, Index from, Index to);
     typedef void (*EnumFn)(EnumOperation, Index, void*&, long&);
